Rebase, upgrade to v3.1.0, use TS in config
