+++ package.json
@@ -13,31 +13,32 @@
     "lint": "eslint --cache \"**/*.{js,jsx,ts,tsx}\""
   },
   "dependencies": {
-    "@docusaurus/core": "3.0.0",
-    "@docusaurus/module-type-aliases": "3.0.0",
-    "@docusaurus/plugin-client-redirects": "3.0.0",
-    "@docusaurus/preset-classic": "3.0.0",
+    "@docusaurus/core": "3.1.0",
+    "@docusaurus/module-type-aliases": "3.1.0",
+    "@docusaurus/plugin-client-redirects": "3.1.0",
+    "@docusaurus/preset-classic": "3.1.0",
+    "@docusaurus/tsconfig": "3.1.0",
+    "@docusaurus/types": "^3.1.0",
     "@mdx-js/react": "^3.0.0",
     "@tailwindcss/typography": "^0.5.10",
-    "@typescript-eslint/eslint-plugin": "^6.12.0",
-    "@typescript-eslint/parser": "^6.12.0",
+    "@typescript-eslint/eslint-plugin": "^6.18.1",
+    "@typescript-eslint/parser": "^6.18.1",
     "autoprefixer": "^10.4.16",
     "cheerio": "^1.0.0-rc.12",
-    "clsx": "^2.0.0",
-    "eslint": "^8.54.0",
+    "clsx": "^2.1.0",
+    "eslint": "^8.56.0",
     "eslint-plugin-jsx-a11y": "^6.8.0",
     "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "postcss": "^8.4.31",
-    "prettier": "^3.1.0",
-    "prism-react-renderer": "^2.3.0",
+    "postcss": "^8.4.33",
+    "prettier": "^3.1.1",
+    "prism-react-renderer": "^2.3.1",
     "raw-loader": "^4.0.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-github-btn": "^1.4.0",
-    "tailwindcss": "^3.3.5",
-    "@docusaurus/tsconfig": "3.0.0",
-    "typescript": "~5.3.2"
+    "tailwindcss": "^3.4.1",
+    "typescript": "~5.3.3"
   },
   "browserslist": {
     "production": [
@@ -56,7 +57,7 @@
   },
   "devDependencies": {
     "husky": "^8.0.3",
-    "lint-staged": "^14.0.1"
+    "lint-staged": "^15.2.0"
   },
   "lint-staged": {
     "**/*": "prettier --write --ignore-unknown"
